1、同步更新Net库内容 2、修复文件传输bug，支持4GB以上文件传输

+++ source/include/FileTransManager.h
@@ -46,7 +46,7 @@ public:
 private:
     FileTransManager();
     bool AddUploadTask(const string &fileid, const string &taskid, const string &filepath, BaseNetWorkSession *session);
-    bool AddDownloadTask(const string &fileid, const string &taskid, const string &filepath, uint32_t filesize, BaseNetWorkSession *session);
+    bool AddDownloadTask(const string &fileid, const string &taskid, const string &filepath, uint64_t filesize, BaseNetWorkSession *session);
     void DeleteTask(const string &taskid);
 
 private:
